F-014: modernize sample app into gold-standard multi-feature structure
Align packageNames and source roots under tools.forma.sample, document the
home/characters feature slice pattern in docs/SAMPLE-APP.md, and keep the
sample application build green.
