+++ application/core/mvvm/library/src/main/java/tools/forma/sample/core/mvvm/library/lifecycle/SingleLiveData.kt
