+++ application/feature/characters/detail/api/build.gradle.kts
@@ -1,5 +1,5 @@
 api(
-    packageName = "tools.forma.sample.character.detail.api",
+    packageName = "tools.forma.sample.feature.characters.detail.api",
     owner = Teams.core,
     dependencies = deps(
         target(":feature:characters:core:api")
