+++ application/feature/characters/detail/res/build.gradle.kts
@@ -1,5 +1,5 @@
 androidRes(
-    packageName = "tools.forma.sample.feature.character.detail.res",
+    packageName = "tools.forma.sample.feature.characters.detail.res",
     dependencies = deps(
         google.material,
         androidx.appcompat,
