+++ application/feature/characters/favorite/api/build.gradle.kts
@@ -1,5 +1,5 @@
 api(
-    packageName = "tools.forma.sample.character.favorite.api",
+    packageName = "tools.forma.sample.feature.characters.favorite.api",
     owner = Teams.core,
     dependencies = deps(
         target(":feature:characters:core:api")
