+++ application/feature/characters/list/api/build.gradle.kts
@@ -1,4 +1,4 @@
 api(
-    packageName = "tools.forma.sample.character.list.api",
+    packageName = "tools.forma.sample.feature.characters.list.api",
     owner = Teams.core,
 )
