+++ application/feature/characters/list/api/src/main/java/tools/forma/sample/feature/characters/list/api/FeatureCharactersListApi.kt
@@ -1,0 +1,7 @@
+package tools.forma.sample.feature.characters.list.api
+
+/**
+ * Public contracts for the characters-list feature.
+ * UI wiring lives in `impl` / `viewbinding`; keep this module dependency-light.
+ */
+object FeatureCharactersListApi
