+++ application/feature/characters/list/res/build.gradle.kts
@@ -1,5 +1,5 @@
 androidRes(
-    packageName = "tools.forma.sample.feature.character.list.res",
+    packageName = "tools.forma.sample.feature.characters.list.res",
     dependencies = deps(
         google.material,
         androidx.appcompat,
