--- application/feature/characters/list/viewbinding/src/main/java/tools/forma/sample/feature/characters/list/impl/domain/model/ICharactersListViewEvent.kt
@@ -1,3 +1,0 @@
-package tools.forma.sample.feature.characters.list.impl.domain.model
-
-interface ICharactersListViewEvent
