--- application/feature/characters/list/viewbinding/src/main/java/tools/forma/sample/feature/characters/list/impl/domain/model/ICharactersListViewState.kt
@@ -1,59 +1,0 @@
-package tools.forma.sample.feature.characters.list.impl.domain.model
-
-interface ICharactersListViewState {
-    /**
-     * Check if current view state is [Refreshing].
-     *
-     * @return True if is refreshing state, otherwise false.
-     */
-    fun isRefreshing(): Boolean
-
-    /**
-     * Check if current view state is [Loaded].
-     *
-     * @return True if is loaded state, otherwise false.
-     */
-    fun isLoaded(): Boolean
-
-    /**
-     * Check if current view state is [Loading].
-     *
-     * @return True if is loading state, otherwise false.
-     */
-    fun isLoading(): Boolean
-
-    /**
-     * Check if current view state is [AddLoading].
-     *
-     * @return True if is add loading state, otherwise false.
-     */
-    fun isAddLoading(): Boolean
-
-    /**
-     * Check if current view state is [Empty].
-     *
-     * @return True if is empty state, otherwise false.
-     */
-    fun isEmpty(): Boolean
-
-    /**
-     * Check if current view state is [Error].
-     *
-     * @return True if is error state, otherwise false.
-     */
-    fun isError(): Boolean
-
-    /**
-     * Check if current view state is [AddError].
-     *
-     * @return True if is add error state, otherwise false.
-     */
-    fun isAddError(): Boolean
-
-    /**
-     * Check if current view state is [NoMoreElements].
-     *
-     * @return True if is no more elements state, otherwise false.
-     */
-    fun isNoMoreElements(): Boolean
-}
