+++ application/feature/characters/list/viewbinding/src/main/java/tools/forma/sample/feature/characters/list/viewbinding/domain/model/ICharactersListViewEvent.kt
@@ -1,0 +1,3 @@
+package tools.forma.sample.feature.characters.list.viewbinding.domain.model
+
+interface ICharactersListViewEvent
