+++ application/feature/home/api/src/main/java/tools/forma/sample/feature/home/api/FeatureHomeApi.kt
@@ -1,0 +1,6 @@
+package tools.forma.sample.feature.home.api
+
+/**
+ * Public contracts for the home shell feature (bottom nav host).
+ */
+object FeatureHomeApi
