--- application/feature/home/api/src/main/java/tools/forma/sample/feature/home/api/empty.kt
@@ -1,2 +1,0 @@
-package tools.forma.sample.feature.home.api
-
