+++ application/feature/home/impl/src/main/java/tools/forma/sample/feature/home/impl/di/HomeModule.kt
@@ -3,7 +3,7 @@ package tools.forma.sample.feature.home.impl.di
 import androidx.lifecycle.ViewModel
 import tools.forma.sample.core.di.library.scopes.FeatureScope
 import tools.forma.sample.core.mvvm.library.di.ViewModelKey
-import tools.forma.sample.feature.home.impl.ui.HomeViewModel
+import tools.forma.sample.feature.home.viewbinding.ui.HomeViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
