+++ application/feature/home/viewbinding/src/main/java/tools/forma/sample/feature/home/viewbinding/ui/HomeViewModel.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2019 vmadalin.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tools.forma.sample.feature.home.viewbinding.ui
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.navigation.NavController
+import tools.forma.sample.feature.home.viewbinding.R
+
+import javax.inject.Inject
+
+// TODO https://github.com/formatools/forma/issues/46
+// Move out from here strong deps of navigation component
+val NAV_FRAGMENTS_ID = setOf(
+    tools.forma.sample.core.navigation.library.R.id.characters_list_fragment,
+    tools.forma.sample.core.navigation.library.R.id.character_favorite_fragment
+)
+
+class HomeViewModel @Inject constructor() : ViewModel() {
+
+    private val _state = MutableLiveData<HomeViewState>()
+    val state: LiveData<HomeViewState>
+        get() = _state
+
+    fun navigationControllerChanged(navController: NavController) {
+        navController.addOnDestinationChangedListener { _, destination, _ ->
+            if (NAV_FRAGMENTS_ID.contains(destination.id)) {
+                _state.postValue(HomeViewState.NavigationScreen)
+            } else {
+                _state.postValue(HomeViewState.FullScreen)
+            }
+        }
+    }
+}
