+++ application/root-app/src/main/java/tools/forma/sample/root/library/SampleMainActivity.kt
@@ -37,6 +37,6 @@ class SampleMainActivity : AppCompatActivity() {
      */
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(tools.forma.sample.rootres.R.layout.activity_main)
+        setContentView(tools.forma.sample.root.res.R.layout.activity_main)
     }
 }
