+++ application/root-res/build.gradle.kts
@@ -1,5 +1,5 @@
 androidRes(
-    packageName = "tools.forma.sample.rootres",
+    packageName = "tools.forma.sample.root.res",
     dependencies = deps(
         androidx.constraintlayout,
         androidx.navigation
