+++ application/toggle-widget/src/main/res/menu/toolbar_menu.xml
@@ -10,7 +10,7 @@
     <item
         android:id="@+id/menu_toggle_theme"
         android:title="@string/home_toolbar_theme"
-        app:actionViewClass="tools.forma.sample.feature.home.impl.ui.menu.ToggleThemeCheckBox"
+        app:actionViewClass="tools.forma.sample.widget.toggle.ToggleThemeCheckBox"
         app:showAsAction="always" />
 
 </menu>
